refactored network code
